+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/BaseBrowserFragment.kt
@@ -39,7 +39,6 @@ import org.mozilla.samples.browser.downloads.DownloadService
 import org.mozilla.samples.browser.ext.components
 import org.mozilla.samples.browser.integration.ContextMenuIntegration
 import org.mozilla.samples.browser.integration.FindInPageIntegration
-import org.mozilla.samples.browser.integration.P2PIntegration
 
 /**
  * Base fragment extended by [BrowserFragment] and [ExternalAppBrowserFragment].
@@ -57,7 +56,6 @@ abstract class BaseBrowserFragment : Fragment(), UserInteractionHandler {
     private val findInPageIntegration = ViewBoundFeatureWrapper<FindInPageIntegration>()
     private val sitePermissionsFeature = ViewBoundFeatureWrapper<SitePermissionsFeature>()
     private val swipeRefreshFeature = ViewBoundFeatureWrapper<SwipeRefreshFeature>()
-    private val p2pIntegration = ViewBoundFeatureWrapper<P2PIntegration>()
 
     protected val sessionId: String?
         get() = arguments?.getString(SESSION_ID_KEY)
@@ -193,21 +191,6 @@ abstract class BaseBrowserFragment : Fragment(), UserInteractionHandler {
             owner = this,
             view = layout)
 
-        p2pIntegration.set(
-            feature = P2PIntegration(
-                store = components.store,
-                engine = components.engine,
-                view = layout.p2p,
-                thunk = { components.nearbyConnection },
-                tabsUseCases = components.tabsUseCases,
-                sessionUseCases = components.sessionUseCases
-            ) { permissions ->
-                requestPermissions(permissions, REQUEST_CODE_P2P_PERMISSIONS)
-            },
-            owner = this,
-            view = layout
-        )
-
         val secureWindowFeature = SecureWindowFeature(
             window = requireActivity().window,
             store = components.store,
@@ -249,7 +232,6 @@ abstract class BaseBrowserFragment : Fragment(), UserInteractionHandler {
             REQUEST_CODE_DOWNLOAD_PERMISSIONS -> downloadsFeature.get()
             REQUEST_CODE_PROMPT_PERMISSIONS -> promptFeature.get()
             REQUEST_CODE_APP_PERMISSIONS -> sitePermissionsFeature.get()
-            REQUEST_CODE_P2P_PERMISSIONS -> p2pIntegration.get()?.feature
             else -> null
         }
         feature?.onPermissionsResult(permissions, grantResults)
@@ -266,7 +248,6 @@ abstract class BaseBrowserFragment : Fragment(), UserInteractionHandler {
         private const val REQUEST_CODE_DOWNLOAD_PERMISSIONS = 1
         private const val REQUEST_CODE_PROMPT_PERMISSIONS = 2
         private const val REQUEST_CODE_APP_PERMISSIONS = 3
-        internal const val REQUEST_CODE_P2P_PERMISSIONS = 4
 
         @JvmStatic
         protected fun Bundle.putSessionId(sessionId: String?) {